Refactor unpack.c for type safety and endian-neutral data access
This commit improves type safety in unpack.c by enforcing explicit type
casting and pointer arithmetic operations. It also introduces endian-neutral
data access methods to ensure that the code can handle data correctly
regardless of the underlying hardware's byte order.

Changes include:
- Explicit casting from void* to specific data types before performing
  arithmetic operations.
- Implementation of endian-neutral functions for reading multi-byte
  numerical values, ensuring consistent behavior across different
  platform architectures.

+++ src/stub.c
@@ -644,7 +644,7 @@ char *SkipArg(char *str)
 /**
    Create a file (OP_CREATE_FILE opcode handler)
 */
-BOOL MakeFile(const char *file_name, DWORD file_size, const void *data)
+BOOL MakeFile(const char *file_name, size_t file_size, const void *data)
 {
     if (file_name == NULL || *file_name == '\0') {
         FATAL("file_name is null or empty");
@@ -674,7 +674,7 @@ BOOL MakeFile(const char *file_name, DWORD file_size, const void *data)
 
         DEBUG("Write data(%lu)", file_size);
 
-        if (!WriteFile(h, data, file_size, &BytesWritten, NULL)) {
+        if (!WriteFile(h, data, (DWORD)file_size, &BytesWritten, NULL)) {
             LAST_ERROR("Write failure");
             result = FALSE;
         }
